refactor: ViewController 코드 분리 - ViewController+Extensions.swift
