+++ WeatherForecast/WeatherForecast/ViewController.swift
@@ -55,30 +55,6 @@ extension ViewController: LocationManagerDelegate {
     }
 }
 
-extension ViewController: UICollectionViewDelegateFlowLayout {
-
-}
-
-extension ViewController: UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 10
-    }
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 1
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: WeatherCollectionViewCell.cellIdentifier, for: indexPath) as? WeatherCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-
-        cell.indexpathLabel.text = "\(indexPath.section + 1)"
-
-        return cell
-    }
-}
-
 class WeatherCollectionViewCell: UICollectionViewCell {
     static let cellIdentifier = "WeatherCollectionViewCell"
